Trait drift_sdk::event_subscriber::EventRpcProvider

pub trait EventRpcProvider: Send + Sync + 'static {
    // Required methods
    fn get_tx_signatures(
        &self,
        account: Pubkey,
        after: Option<Signature>,
        limit: Option<usize>
    ) -> BoxFuture<'_, SdkResult<Vec<String>>>;
    fn get_tx(
        &self,
        signature: Signature
    ) -> BoxFuture<'_, SdkResult<EncodedTransactionWithStatusMeta>>;
}

RPC functions required for drift event subscriptions

Required Methods

fn get_tx_signatures( &self, account: Pubkey, after: Option<Signature>, limit: Option<usize> ) -> BoxFuture<'_, SdkResult<Vec<String>>>

Fetch tx signatures of account after only return txs more recent than this signature, if given limit return at most this many signatures, if given

fn get_tx( &self, signature: Signature ) -> BoxFuture<'_, SdkResult<EncodedTransactionWithStatusMeta>>

Fetch tx with signature