Struct drift_sdk::DriftClient

pub struct DriftClient<T: AccountProvider> {
    pub active_sub_account_id: u8,
    pub sub_account_ids: Vec<u8>,
    /* private fields */
}

Drift Client API

It is cheaply clone-able and consumers are encouraged to do so It is not recommended to create multiple instances with ::new() as this will not re-use underlying resources such as network connections or memory allocations

Fields
active_sub_account_id: u8sub_account_ids: Vec<u8>
Implementations

impl<T: AccountProvider> DriftClient<T>

pub async fn new( context: Context, account_provider: T, keypair: Keypair ) -> SdkResult<Self>

pub async fn new_with_opts( context: Context, account_provider: T, keypair: Keypair, opts: ClientOpts ) -> SdkResult<Self>

pub fn inner(&self) -> &RpcClient

Return a handle to the inner RPC client

pub fn program_data(&self) -> &ProgramData

Return on-chain program metadata

pub fn get_sub_account_id_for_ix(&self, sub_account_id: Option<u8>) -> u8

Get the active sub account id

pub async fn get_order_by_id( &self, account: &Pubkey, order_id: u32 ) -> SdkResult<Option<Order>>

Get an account’s open order by id

account the drift user PDA

pub async fn get_order_by_user_id( &self, account: &Pubkey, user_order_id: u8 ) -> SdkResult<Option<Order>>

Get an account’s open order by user assigned id

account the drift user PDA

pub async fn all_orders(&self, account: &Pubkey) -> SdkResult<Vec<Order>>

Get all the account’s open orders

account the drift user PDA

pub async fn all_positions( &self, account: &Pubkey ) -> SdkResult<(Vec<SpotPosition>, Vec<PerpPosition>)>

Get all the account’s active positions

account the drift user PDA

pub async fn perp_position( &self, account: &Pubkey, market_index: u16 ) -> SdkResult<Option<PerpPosition>>

Get a perp position by market

account the drift user PDA

Returns the position if it exists

pub async fn spot_position( &self, account: &Pubkey, market_index: u16 ) -> SdkResult<Option<SpotPosition>>

Get a spot position by market

account the drift user PDA

Returns the position if it exists

pub fn wallet(&self) -> &Wallet

Return the DriftClient’s wallet

pub async fn get_user_account(&self, account: &Pubkey) -> SdkResult<User>

Get the user account data

account the drift user PDA

Returns the deserialized account data (User)

pub async fn sign_and_send(&self, tx: VersionedMessage) -> SdkResult<Signature>

Sign and send a tx to the network

Returns the signature on success

pub async fn sign_and_send_with_config( &self, tx: VersionedMessage, config: RpcSendTransactionConfig ) -> SdkResult<Signature>

Sign and send a tx to the network

Returns the signature on success

pub async fn get_spot_market_info( &self, market_index: u16 ) -> SdkResult<SpotMarket>

Get live info of a spot market

pub fn market_lookup(&self, symbol: &str) -> Option<MarketId>

Lookup a market by symbol

This operation is not free so lookups should be reused/cached by the caller

Returns None if symbol does not map to any known market

pub async fn oracle_price(&self, market: MarketId) -> SdkResult<i64>

Get live oracle price for market

pub async fn init_tx( &self, account: &Pubkey, delegated: bool ) -> SdkResult<TransactionBuilder<'_>>

Initialize a transaction given a (sub)account address

let tx = client
    .init_tx(&wallet.sub_account(3), false)
    .cancel_all_orders()
    .place_orders(...)
    .build();

Returns a TransactionBuilder for composing the tx

pub async fn get_recent_priority_fees( &self, writable_markets: &[MarketId], window: Option<usize> ) -> SdkResult<Vec<u64>>