Struct drift_sdk::Wallet

pub struct Wallet { /* private fields */ }

Drift wallet

Implementations

impl Wallet

pub fn is_delegated(&self) -> bool

Returns true if the wallet is configured for delegated signing

pub fn try_from_str(path_or_key: &str) -> SdkResult<Self>

Init wallet from a string that could be either a file path or the encoded key, uses default sub-account

pub fn read_only(authority: Pubkey) -> Self

Construct a read-only wallet

pub fn from_seed_bs58(seed: &str) -> Self

Init wallet from base58 encoded seed, uses default sub-account

panics

if the key is invalid

pub fn from_seed(seed: &[u8]) -> SdkResult<Self>

Init wallet from seed bytes, uses default sub-account

pub fn new(authority: Keypair) -> Self

Init wallet with keypair

authority keypair for tx signing

pub fn to_delegated(&mut self, authority: Pubkey)

Convert the wallet into a delegated one by providing the authority public key

pub fn derive_user_account( account: &Pubkey, sub_account_id: u16, program: &Pubkey ) -> Pubkey

Calculate the address of a drift user account/sub-account

pub fn derive_stats_account(account: &Pubkey, program: &Pubkey) -> Pubkey

Calculate the address of a drift stats account

pub fn sign_tx( &self, message: VersionedMessage, recent_block_hash: Hash ) -> SdkResult<VersionedTransaction>

Signs the given tx message returning the tx on success

pub fn authority(&self) -> &Pubkey

Return the wallet authority address

pub fn signer(&self) -> Pubkey

Return the wallet signing address

pub fn stats(&self) -> &Pubkey

Return the drift user stats address

pub fn default_sub_account(&self) -> Pubkey

Return the address of the default sub-account (0)

pub fn sub_account(&self, sub_account_id: u16) -> Pubkey

Calculate the drift user address given a sub_account_id