Struct drift_sdk::types::Order

#[repr(C)]
pub struct Order {

    pub slot: u64,
    pub price: u64,
    pub base_asset_amount: u64,
    pub base_asset_amount_filled: u64,
    pub quote_asset_amount_filled: u64,
    pub trigger_price: u64,
    pub auction_start_price: i64,
    pub auction_end_price: i64,
    pub max_ts: i64,
    pub oracle_price_offset: i32,
    pub order_id: u32,
    pub market_index: u16,
    pub status: OrderStatus,
    pub order_type: OrderType,
    pub market_type: MarketType,
    pub user_order_id: u8,
    pub existing_position_direction: PositionDirection,
    pub direction: PositionDirection,
    pub reduce_only: bool,
    pub post_only: bool,
    pub immediate_or_cancel: bool,
    pub trigger_condition: OrderTriggerCondition,
    pub auction_duration: u8,
    pub padding: [u8; 3],

}

Fields
slot: u64

The slot the order was placed

price: u64

The limit price for the order (can be 0 for market orders) For orders with an auction, this price isn’t used until the auction is complete precision: PRICE_PRECISION

base_asset_amount: u64

The size of the order precision for perps: BASE_PRECISION precision for spot: token mint precision

base_asset_amount_filled: u64

The amount of the order filled precision for perps: BASE_PRECISION precision for spot: token mint precision

quote_asset_amount_filled: u64

The amount of quote filled for the order precision: QUOTE_PRECISION

trigger_price: u64

At what price the order will be triggered. Only relevant for trigger orders precision: PRICE_PRECISION

auction_start_price: i64

The start price for the auction. Only relevant for market/oracle orders precision: PRICE_PRECISION

auction_end_price: i64

The end price for the auction. Only relevant for market/oracle orders precision: PRICE_PRECISION

max_ts: i64

The time when the order will expire

oracle_price_offset: i32

If set, the order limit price is the oracle price + this offset precision: PRICE_PRECISION

order_id: u32

The id for the order. Each users has their own order id space

market_index: u16

The perp/spot market index

status: OrderStatus

Whether the order is open or unused

order_type: OrderType

The type of order

market_type: MarketType

Whether market is spot or perp

user_order_id: u8

User generated order id. Can make it easier to place/cancel orders

existing_position_direction: PositionDirection

What the users position was when the order was placed

direction: PositionDirection

Whether the user is going long or short. LONG = bid, SHORT = ask

reduce_only: bool

Whether the order is allowed to only reduce position size

post_only: bool

Whether the order must be a maker

immediate_or_cancel: bool

Whether the order must be canceled the same slot it is placed

trigger_condition: OrderTriggerCondition

Whether the order is triggered above or below the trigger price. Only relevant for trigger orders

auction_duration: u8

How many slots the auction lasts

padding: [u8; 3]
Implementations

impl Order

pub fn seconds_til_expiry(self, now: i64) -> i64

pub fn has_oracle_price_offset(self) -> bool

pub fn get_limit_price( &self, valid_oracle_price: Option<i64>, fallback_price: Option<u64>, slot: u64, tick_size: u64 ) -> Result<Option<u64>, ErrorCode>

pub fn force_get_limit_price( &self, valid_oracle_price: Option<i64>, fallback_price: Option<u64>, slot: u64, tick_size: u64 ) -> Result<u64, ErrorCode>

pub fn has_limit_price(self, slot: u64) -> Result<bool, ErrorCode>

pub fn is_auction_complete(self, slot: u64) -> Result<bool, ErrorCode>

pub fn has_auction(&self) -> bool

pub fn has_auction_price( &self, order_slot: u64, auction_duration: u8, slot: u64 ) -> Result<bool, ErrorCode>

pub fn get_base_asset_amount_unfilled( &self, existing_position: Option<i64> ) -> Result<u64, ErrorCode>

Passing in an existing_position forces the function to consider the order’s reduce only status

pub fn get_standardized_base_asset_amount_unfilled( &self, existing_position: Option<i64>, step_size: u64 ) -> Result<u64, ErrorCode>

Stardardizes the base asset amount unfilled to the nearest step size Particularly important for spot positions where existing position can be dust

pub fn must_be_triggered(&self) -> bool

pub fn triggered(&self) -> bool

pub fn is_jit_maker(&self) -> bool

pub fn is_open_order_for_market( &self, market_index: u16, market_type: &MarketType ) -> bool

pub fn get_spot_position_update_direction( &self, asset_type: AssetType ) -> SpotBalanceType

pub fn is_market_order(&self) -> bool

pub fn is_limit_order(&self) -> bool

pub fn is_resting_limit_order(&self, slot: u64) -> Result<bool, ErrorCode>