Struct drift_sdk::types::SpotMarket

#[repr(C)]
pub struct SpotMarket {

    pub pubkey: Pubkey,
    pub oracle: Pubkey,
    pub mint: Pubkey,
    pub vault: Pubkey,
    pub name: [u8; 32],
    pub historical_oracle_data: HistoricalOracleData,
    pub historical_index_data: HistoricalIndexData,
    pub revenue_pool: PoolBalance,
    pub spot_fee_pool: PoolBalance,
    pub insurance_fund: InsuranceFund,
    pub total_spot_fee: u128,
    pub deposit_balance: u128,
    pub borrow_balance: u128,
    pub cumulative_deposit_interest: u128,
    pub cumulative_borrow_interest: u128,
    pub total_social_loss: u128,
    pub total_quote_social_loss: u128,
    pub withdraw_guard_threshold: u64,
    pub max_token_deposits: u64,
    pub deposit_token_twap: u64,
    pub borrow_token_twap: u64,
    pub utilization_twap: u64,
    pub last_interest_ts: u64,
    pub last_twap_ts: u64,
    pub expiry_ts: i64,
    pub order_step_size: u64,
    pub order_tick_size: u64,
    pub min_order_size: u64,
    pub max_position_size: u64,
    pub next_fill_record_id: u64,
    pub next_deposit_record_id: u64,
    pub initial_asset_weight: u32,
    pub maintenance_asset_weight: u32,
    pub initial_liability_weight: u32,
    pub maintenance_liability_weight: u32,
    pub imf_factor: u32,
    pub liquidator_fee: u32,
    pub if_liquidation_fee: u32,
    pub optimal_utilization: u32,
    pub optimal_borrow_rate: u32,
    pub max_borrow_rate: u32,
    pub decimals: u32,
    pub market_index: u16,
    pub orders_enabled: bool,
    pub oracle_source: OracleSource,
    pub status: MarketStatus,
    pub asset_tier: AssetTier,
    pub padding1: [u8; 6],
    pub flash_loan_amount: u64,
    pub flash_loan_initial_token_amount: u64,
    pub total_swap_fee: u64,
    pub scale_initial_asset_weight_start: u64,
    pub padding: [u8; 48],

}

Fields
pubkey: Pubkey

The address of the spot market. It is a pda of the market index

oracle: Pubkey

The oracle used to price the markets deposits/borrows

mint: Pubkey

The token mint of the market

vault: Pubkey

The vault used to store the market’s deposits The amount in the vault should be equal to or greater than deposits - borrows

name: [u8; 32]

The encoded display name for the market e.g. SOL

historical_oracle_data: HistoricalOracleDatahistorical_index_data: HistoricalIndexDatarevenue_pool: PoolBalance

Revenue the protocol has collected in this markets token e.g. for SOL-PERP, funds can be settled in usdc and will flow into the USDC revenue pool

spot_fee_pool: PoolBalance

The fees collected from swaps between this market and the quote market Is settled to the quote markets revenue pool

insurance_fund: InsuranceFund

Details on the insurance fund covering bankruptcies in this markets token Covers bankruptcies for borrows with this markets token and perps settling in this markets token

total_spot_fee: u128

The total spot fees collected for this market precision: QUOTE_PRECISION

deposit_balance: u128

The sum of the scaled balances for deposits across users and pool balances To convert to the deposit token amount, multiply by the cumulative deposit interest precision: SPOT_BALANCE_PRECISION

borrow_balance: u128

The sum of the scaled balances for borrows across users and pool balances To convert to the borrow token amount, multiply by the cumulative borrow interest precision: SPOT_BALANCE_PRECISION

cumulative_deposit_interest: u128

The cumulative interest earned by depositors Used to calculate the deposit token amount from the deposit balance precision: SPOT_CUMULATIVE_INTEREST_PRECISION

cumulative_borrow_interest: u128

The cumulative interest earned by borrowers Used to calculate the borrow token amount from the borrow balance precision: SPOT_CUMULATIVE_INTEREST_PRECISION

total_social_loss: u128

The total socialized loss from borrows, in the mint’s token precision: token mint precision

total_quote_social_loss: u128

The total socialized loss from borrows, in the quote market’s token preicision: QUOTE_PRECISION

withdraw_guard_threshold: u64

no withdraw limits/guards when deposits below this threshold precision: token mint precision

max_token_deposits: u64

The max amount of token deposits in this market 0 if there is no limit precision: token mint precision

deposit_token_twap: u64

24hr average of deposit token amount precision: token mint precision

borrow_token_twap: u64

24hr average of borrow token amount precision: token mint precision

utilization_twap: u64

24hr average of utilization which is borrow amount over token amount precision: SPOT_UTILIZATION_PRECISION

last_interest_ts: u64

Last time the cumulative deposit and borrow interest was updated

last_twap_ts: u64

Last time the deposit/borrow/utilization averages were updated

expiry_ts: i64

The time the market is set to expire. Only set if market is in reduce only mode

order_step_size: u64

Spot orders must be a multiple of the step size precision: token mint precision

order_tick_size: u64

Spot orders must be a multiple of the tick size precision: PRICE_PRECISION

min_order_size: u64

The minimum order size precision: token mint precision

max_position_size: u64

The maximum spot position size if the limit is 0, there is no limit precision: token mint precision

next_fill_record_id: u64

Every spot trade has a fill record id. This is the next id to use

next_deposit_record_id: u64

Every deposit has a deposit record id. This is the next id to use

initial_asset_weight: u32

The initial asset weight used to calculate a deposits contribution to a users initial total collateral e.g. if the asset weight is .8, $100 of deposits contributes $80 to the users initial total collateral precision: SPOT_WEIGHT_PRECISION

maintenance_asset_weight: u32

The maintenance asset weight used to calculate a deposits contribution to a users maintenance total collateral e.g. if the asset weight is .9, $100 of deposits contributes $90 to the users maintenance total collateral precision: SPOT_WEIGHT_PRECISION

initial_liability_weight: u32

The initial liability weight used to calculate a borrows contribution to a users initial margin requirement e.g. if the liability weight is .9, $100 of borrows contributes $90 to the users initial margin requirement precision: SPOT_WEIGHT_PRECISION

maintenance_liability_weight: u32

The maintenance liability weight used to calculate a borrows contribution to a users maintenance margin requirement e.g. if the liability weight is .8, $100 of borrows contributes $80 to the users maintenance margin requirement precision: SPOT_WEIGHT_PRECISION

imf_factor: u32

The initial margin fraction factor. Used to increase liability weight/decrease asset weight for large positions precision: MARGIN_PRECISION

liquidator_fee: u32

The fee the liquidator is paid for taking over borrow/deposit precision: LIQUIDATOR_FEE_PRECISION

if_liquidation_fee: u32

The fee the insurance fund receives from liquidation precision: LIQUIDATOR_FEE_PRECISION

optimal_utilization: u32

The optimal utilization rate for this market. Used to determine the markets borrow rate precision: SPOT_UTILIZATION_PRECISION

optimal_borrow_rate: u32

The borrow rate for this market when the market has optimal utilization precision: SPOT_RATE_PRECISION

max_borrow_rate: u32

The borrow rate for this market when the market has 1000 utilization precision: SPOT_RATE_PRECISION

decimals: u32

The market’s token mint’s decimals. To from decimals to a precision, 10^decimals

market_index: u16orders_enabled: bool

Whether or not spot trading is enabled

oracle_source: OracleSourcestatus: MarketStatusasset_tier: AssetTier

The asset tier affects how a deposit can be used as collateral and the priority for a borrow being liquidated

padding1: [u8; 6]flash_loan_amount: u64

For swaps, the amount of token loaned out in the begin_swap ix precision: token mint precision

flash_loan_initial_token_amount: u64

For swaps, the amount in the users token account in the begin_swap ix Used to calculate how much of the token left the system in end_swap ix precision: token mint precision

total_swap_fee: u64

The total fees received from swaps precision: token mint precision

scale_initial_asset_weight_start: u64

When to begin scaling down the initial asset weight disabled when 0 precision: QUOTE_PRECISION

padding: [u8; 48]
Implementations

impl SpotMarket

pub fn is_active(&self, now: i64) -> Result<bool, ErrorCode>

pub fn is_reduce_only(&self) -> bool

pub fn fills_enabled(&self) -> bool

pub fn get_sanitize_clamp_denominator(&self) -> Result<Option<i64>, ErrorCode>

pub fn get_asset_weight( &self, size: u128, oracle_price: i64, margin_requirement_type: &MarginRequirementType ) -> Result<u32, ErrorCode>

pub fn get_scaled_initial_asset_weight( &self, oracle_price: i64 ) -> Result<u32, ErrorCode>

pub fn get_liability_weight( &self, size: u128, margin_requirement_type: &MarginRequirementType ) -> Result<u32, ErrorCode>

pub fn get_margin_ratio( &self, margin_requirement_type: &MarginRequirementType ) -> Result<u32, ErrorCode>

pub fn get_deposits(&self) -> Result<u128, ErrorCode>

pub fn get_borrows(&self) -> Result<u128, ErrorCode>

pub fn validate_max_token_deposits(&self) -> Result<(), ErrorCode>

pub fn get_available_deposits(&self) -> Result<u128, ErrorCode>

pub fn get_precision(self) -> u64

pub fn get_utilization(self) -> Result<u128, ErrorCode>

pub fn is_healthy_utilization(self) -> Result<bool, ErrorCode>