Struct drift_sdk::types::SpotPosition

#[repr(C)]
pub struct SpotPosition {
    pub scaled_balance: u64,
    pub open_bids: i64,
    pub open_asks: i64,
    pub cumulative_deposits: i64,
    pub market_index: u16,
    pub balance_type: SpotBalanceType,
    pub open_orders: u8,
    pub padding: [u8; 4],
}

Fields
scaled_balance: u64

The scaled balance of the position. To get the token amount, multiply by the cumulative deposit/borrow interest of corresponding market. precision: SPOT_BALANCE_PRECISION

open_bids: i64

How many spot bids the user has open precision: token mint precision

open_asks: i64

How many spot asks the user has open precision: token mint precision

cumulative_deposits: i64

The cumulative deposits/borrows a user has made into a market precision: token mint precision

market_index: u16

The market index of the corresponding spot market

balance_type: SpotBalanceType

Whether the position is deposit or borrow

open_orders: u8

Number of open orders

padding: [u8; 4]
Implementations

impl SpotPosition

pub fn is_available(&self) -> bool

pub fn has_open_order(&self) -> bool

pub fn margin_requirement_for_open_orders(&self) -> Result<u128, ErrorCode>

pub fn get_token_amount( &self, spot_market: &SpotMarket ) -> Result<u128, ErrorCode>

pub fn get_signed_token_amount( &self, spot_market: &SpotMarket ) -> Result<i128, ErrorCode>

pub fn get_worst_case_fill_simulation( &self, spot_market: &SpotMarket, strict_oracle_price: &StrictOraclePrice, token_amount: Option<i128>, margin_type: MarginRequirementType ) -> Result<OrderFillSimulation, ErrorCode>

pub fn simulate_fills_both_sides( &self, spot_market: &SpotMarket, strict_oracle_price: &StrictOraclePrice, token_amount: Option<i128>, margin_type: MarginRequirementType ) -> Result<[OrderFillSimulation; 2], ErrorCode>

pub fn is_borrow(&self) -> bool